fix: the page 'take an exercise' only shows the exercises where status is 'answering'
