--- src/models/exercice.php
@@ -1,84 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use App\Controllers\Database;
-
-class Exercice
-{
-    public $id;
-    public $title;
-    public $status;
-
-    protected function __construct($id, $title, $status)
-    {
-        $this->id = $id;
-        $this->title = $title;
-        $this->status = $status;
-    }
-
-    public static function getAllExercices()
-    {
-        $db = new Database();
-
-        $results = $db->executeQuery("SELECT id, title, status FROM exercices");
-
-        $return = [];
-
-        foreach ($results as $result) {
-            $return[] = new Exercice($result['id'], $result['title'], $result['status']);
-        }
-        return $return;
-    }
-
-    public static function createExercice($title)
-    {
-        $db = new Database();
-
-        $db->executeQuery("INSERT INTO exercices (title) VALUES ('{$title}')");
-    }
-
-    public static function getExercicesByStatus($status)
-    {
-        $db = new Database();
-
-        $results = $db->executeQuery("SELECT id, title, status FROM exercices WHERE status = {$status}");
-
-        $return = [];
-
-        foreach ($results as $result) {
-            $return[] = new Exercice($result['id'], $result['title'], $result['status']);
-        }
-
-        return $return;
-    }
-
-    public static function getExerciceById($id)
-    {
-        $db = new Database();
-
-        $results = $db->executeQuery("SELECT id, title, status FROM exercices WHERE id = {$id}");
-
-        $return = [];
-
-        foreach ($results as $result) {
-            $return[] = new Exercice($result['id'], $result['title'], $result['status']);
-        }
-
-        return $return;
-    }
-
-    public static function updateExercice($fields, $id)
-    {
-        $db = new Database();
-
-        $querybuilder = '';
-
-        foreach ($fields as $field)
-        {
-            $querybuilder = "{fields[]} = {$field}";
-        }
-
-        $db->executeQuery("UPDATE exercices SET {$querybuilder} WHERE id = {$id}");
-    }
-}
